Pass user model into LandingPage.State.update
I'm not sure this is the right solution, but it seems like the most
simple one to solve the problem of affecting something outside of the
main parameter model.

An alternative would be for the feature component to cause some "global
effect" that the App would pick up on and react to it by triggering
a user action that causes it to reload. I prefer this way because I feel
like I can reuse it easier if I just respect the parameter requirements.