Store jwt token && check api/current_user on init
This commit is doing quite a bit so I'm going to explain it since I've
had a few glasses of wine at the time I'm committing.

Users who have signed in before will have a jwtToken[1] which will
get passed as an application flag in Elm bootup. (inside index.js) When
Elm bootstraps the startup, it will check if you have a jwtToken. If you
do, it will make a request to /current_user which will provide the
rest of the signed in user data.

As of the timestamp on this commit, there are no good or even really
plausible libraries I could use so I had to use some js localStorage
library and interop with it in Elm. The resulting code is a little messy
but I don't suspect a refactor will be very expensive.

[1] At the time of this commit, the "jwtToken" isn't getting written
anywhere, I'll leave that for a future commit when I'm not so sleepy.